TP casi terminado, queda ver de agregar alguna mejora

+++ tp1/src/funciones.c
@@ -1,5 +1,13 @@
-//a) Calcular la suma (A+B)
 
+/**a) Calcular la suma (A+B)
+ * Sumar: realiza la suma de dos numeros float
+ *
+ * par1: 1er numero float
+ * par2: 2do numero float
+ *
+ * return: la suma en numero float
+ *
+ */
 float Sumar (float numero_uno, float numero_dos)
 {
 	float suma;
@@ -7,8 +15,17 @@ float Sumar (float numero_uno, float numero_dos)
 	return suma;
 }
 
+//------------------------------------------------------------------------
 
-//b) Calcular la resta (A-B)
+/**b) Calcular la resta (A-B)
+ * Restar: realiza la resta de dos numeros float
+ *
+ * par1: 1er numero float
+ * par2: 2do numero float
+ *
+ * return: la resta en numero float
+ *
+ */
 float Restar (float numero_uno, float numero_dos)
 {
 	float resta;
@@ -16,7 +33,18 @@ float Restar (float numero_uno, float numero_dos)
 	return resta;
 }
 
-//c) Calcular la division (A/B)
+//------------------------------------------------------------------------
+
+//
+/**c) Calcular la division (A/B)
+ * Dividir: realiza la division de dos numeros float
+ *
+ * par1: 1er numero float (dividendo)
+ * par2: 2do numero float (divisor)
+ *
+ * return: la division en numero float
+ *
+ */
 float Dividir (float numero_uno, float numero_dos)
 {
 	float division;
@@ -24,7 +52,18 @@ float Dividir (float numero_uno, float numero_dos)
 	return division;
 }
 
-// d) Calcular la multiplicacion (A*B)
+//------------------------------------------------------------------------
+
+
+/**d) Calcular la multiplicacion (A*B)
+ * Multiplicar: realiza la multiplicacion de dos numeros float
+ *
+ * par1: 1er numero float
+ * par2: 2do numero float
+ *
+ * return: el producto en numero DOUBLE
+ *
+ */
 double Multiplicar (float numero_uno, float numero_dos)
 {
 	double multiplicacion;
@@ -33,10 +72,21 @@ double Multiplicar (float numero_uno, float numero_dos)
 }
 
 
+//------------------------------------------------------------------------
+
+
+/**e) Calcular el factorial (A!)
+ * Factorial: realiza el factorial de UN NUMERO INT
+ *
+ * par1: UN numero INT
+ *
+ * return: el FACTORIAL en numero DOUBLE
+ *
+ */
 double Factorial (int numero)
 {
-	int i;
-	int fractorial=1;
+	double i;
+	double fractorial=1;
 
 	for (i=numero; i>0; i--)
 	{
